new pic

+++ client/src/PharmaProvider.js
@@ -12,7 +12,6 @@ class PharmaProvider extends Component {
     constructor(){
         super()
         this.state = {
-            data:[],
             toggleHome: true,
             toggle3: '',
             user: JSON.parse(localStorage.getItem("user")) || {},
@@ -73,19 +72,17 @@ class PharmaProvider extends Component {
 
     login = userInfo => {
         axios.post('/user/login', userInfo).then(res => {
-            this.setState({data: res.data})
-            const { token, user } = res.data          // when the token and user comes back from the database we store it in local storage
-            localStorage.setItem("user", JSON.stringify(user))
-            localStorage.setItem("token", token)
-            this.setState({ user: user, token })
-        }).catch(err => this.handleLoginError(err, this.state.data))
-    }
-
-    handleLoginError = (err, data) => {
-        this.setState({alert: err.response.data.errMsg}) 
-        if(err.response.data.errMsg === "Vã rugãm sã activați contul!"){
-            this.sendActivationEmail(data.user._id, data.user.username)
-        }
+            console.log(res.data)
+            if(res.data.confirmed){
+                const { token, user } = res.data          // when the token and user comes back from the database we store it in local storage
+                localStorage.setItem("user", JSON.stringify(user))
+                localStorage.setItem("token", token)
+                this.setState({ user: user, token })
+            }else{
+                this.setState({alert: "Vã rugãm sã activati contul!"})
+                this.sendActivationEmail(res.data._id, res.data.username)
+            }
+        }).catch(err => this.setState({alert: err.response.data.errMsg}))
     }
 
     editToggler = () => {
